Move `PreferenceTable` creation to its own model creator
Till now it has lived in pkg `load`, but is more logical to have it be a
standalone model with its own creator.

As a byproduct of this change, the existing `load.LoadFromIO` function
which referenced `load.LoadFromDataType` is also modified to return the
loaded JSON instead of a fully created `PreferenceTable`.

Scope:

* Move `load.LoadFromDataType` -> `core.NewPreferenceTable`
* Update `load.LoadFromIO` to return `*[]core.MatchEntry`